requirements.txt
langgraph 
langchain 
langchain-openai 
openai
python-dotenv
beautifulsoup4
pypdf
python-docx
qdrant-client
fastmcp
telethon
boto3
playwright
